Removed pre go112 build code
