+++ builder_go_module_test.go
@@ -1,5 +1,3 @@
-// +build go1.11
-
 package godog
 
 import (
